debug physical test #15.2.2

+++ ports/nrf52840/core/hardware/ads1015/ads1015.hpp
@@ -58,5 +58,5 @@ public:
 
 private:
 ads1015LL m_ads1015;
-int m_digital_value; 
+double m_digital_value; 
 };
